pyproject.toml
[project]
name = "viewpoint-estimator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "huggingface-hub>=0.36.0",
    "jupyterlab>=4.5.0",
    "numpy>=2.1.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "xformers>=0.0.33.post1",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }
